+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:55612/jest?","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:55669/jest?","mongoDBName":"jest"}
