+++ test/login-validation.spec.ts
@@ -1,8 +1,6 @@
 import { makeLoginValidation } from '../src/main/factories/login/login-validation'
-import { EmailValidation } from '../src/presentation/helpers/validators/email-validation'
-import { RequiredFieldValidation } from '../src/presentation/helpers/validators/required-field-validation'
+import { EmailValidation, RequiredFieldValidation, ValidationComposite } from '../src/presentation/helpers/validators'
 import { Validation } from '../src/presentation/protocols/validation'
-import { ValidationComposite } from '../src/presentation/helpers/validators/validation-composite'
 import { EmailValidator } from '../src/presentation/protocols/email-validator'
 
 jest.mock('../src/presentation/helpers/validators/validation-composite')
