+++ test/signup-validation.spec.ts
@@ -1,9 +1,6 @@
 import { makeSignUpValidation } from '../src/main/factories/signup/signup-validation'
-import { CompareFieldsValidation } from '../src/presentation/helpers/validators/compare-fields-validation'
-import { EmailValidation } from '../src/presentation/helpers/validators/email-validation'
-import { RequiredFieldValidation } from '../src/presentation/helpers/validators/required-field-validation'
+import { CompareFieldsValidation, EmailValidation, RequiredFieldValidation, ValidationComposite } from '../src/presentation/helpers/validators'
 import { Validation } from '../src/presentation/protocols/validation'
-import { ValidationComposite } from '../src/presentation/helpers/validators/validation-composite'
 import { EmailValidator } from '../src/presentation/protocols/email-validator'
 
 jest.mock('../src/presentation/helpers/validators/validation-composite')
